Checks out the branch that is deployed
I have setup `ms-nlu` repository to deploy automatically on successful builds on stable branch. 

Currently the builds are seeming to fail at this step. Refer https://github.com/onboardiq/ms-nlu/actions/runs/4852685252/jobs/8654833163

I wonder if this is the fix needed to make sure the code is able to checkout the SHA that is actually being deployed